Move peermanager to own file
